Fix possible portability problem, per buildfarm warnings.

+++ contrib/pgbench/pgbench.c
@@ -1,5 +1,5 @@
 /*
- * $PostgreSQL: pgsql/contrib/pgbench/pgbench.c,v 1.68 2007/07/06 20:17:02 wieck Exp $
+ * $PostgreSQL: pgsql/contrib/pgbench/pgbench.c,v 1.69 2007/07/15 22:34:26 tgl Exp $
  *
  * pgbench: a simple benchmark program for PostgreSQL
  * written by Tatsuo Ishii
@@ -1442,7 +1442,7 @@ main(int argc, char **argv)
 	{
 		char		logpath[64];
 
-		snprintf(logpath, 64, "pgbench_log.%d", getpid());
+		snprintf(logpath, 64, "pgbench_log.%d", (int) getpid());
 		LOGFILE = fopen(logpath, "w");
 
 		if (LOGFILE == NULL)
